moved several authority related functions and protocol buffers definition to subp2p-explorer as shared utilities
