+++ subp2p-explorer/Cargo.toml
@@ -34,4 +34,9 @@ primitive-types = { workspace = true, default-features = false, features = ["cod
 hex = { workspace = true }
 schnorrkel = "0.11.4"
 blake2 = "0.10.4"
+multihash-codetable = { workspace = true, features = ["digest", "serde", "sha2"] }
 bs58 = { version = "0.5.0", features = ["alloc"] }
+prost = "0.12"
+
+[build-dependencies]
+prost-build = "0.11"
