+++ subp2p-explorer/src/util/mod.rs
@@ -2,5 +2,7 @@
 // This file is dual-licensed as Apache-2.0 or GPL-3.0.
 // see LICENSE for license details.
 
+pub mod authorities;
 pub mod crypto;
+pub mod p2p;
 pub mod ss58;
